pyproject.toml
[tool.poetry]
name = "dop-be"
version = "0.1.0"
description = ""
authors = ["nvdluan <nvdluan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^1.3.20"
psycopg2-binary = "^2.8.5"
fastapi-utils = "^0.2.1"
alembic = "^1.7.6"
python-dotenv = "^0.19.2"
pytz = "^2022.1"
pycryptodome = "^3.15.0"
Pyrebase = "3.0.1"
PyJWT = "1.7.1"
schedule = "^1.1.0"
tldextract = "^3.4.0"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
pusher = "^3.3.2"
python-multipart = "^0.0.6"
jinja2 = "3.0.3"
aiofiles = "^23.1.0"
pillow = "^9.5.0"
pydot ="^1.4.2"
pandas="^2.0.3"
httpx="^0.24.1"
gunicorn = "^21.2.0"
openpyxl = "^3.1.2"
boto3= "^1.28.73"
[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
